+++ rdsai-20250507/alibabacloud_rdsai20250507/models/_create_custom_agent_shrink_request.py
@@ -9,6 +9,7 @@ class CreateCustomAgentShrinkRequest(DaraModel):
         self,
         enable_tools: bool = None,
         name: str = None,
+        skill_ids_shrink: str = None,
         system_prompt: str = None,
         tools_shrink: str = None,
     ):
@@ -16,6 +17,7 @@ class CreateCustomAgentShrinkRequest(DaraModel):
         self.enable_tools = enable_tools
         # The operation that you want to perform. Set the value to **CreateCustomAgent**.
         self.name = name
+        self.skill_ids_shrink = skill_ids_shrink
         # The name of the dedicated agent.
         # 
         # This parameter is required.
@@ -37,6 +39,9 @@ class CreateCustomAgentShrinkRequest(DaraModel):
         if self.name is not None:
             result['Name'] = self.name
 
+        if self.skill_ids_shrink is not None:
+            result['SkillIds'] = self.skill_ids_shrink
+
         if self.system_prompt is not None:
             result['SystemPrompt'] = self.system_prompt
 
@@ -52,6 +57,9 @@ class CreateCustomAgentShrinkRequest(DaraModel):
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
+
+        if m.get('SkillIds') is not None:
+            self.skill_ids_shrink = m.get('SkillIds')
 
         if m.get('SystemPrompt') is not None:
             self.system_prompt = m.get('SystemPrompt')