+++ rdsai-20250507/alibabacloud_rdsai20250507/models/_list_custom_agent_response_body.py
@@ -14,6 +14,7 @@ class ListCustomAgentResponseBody(DaraModel):
         page_number: int = None,
         page_size: int = None,
         request_id: str = None,
+        skills: List[main_models.ListCustomAgentResponseBodySkills] = None,
         total_count: int = None,
     ):
         # The returned data.
@@ -24,12 +25,17 @@ class ListCustomAgentResponseBody(DaraModel):
         self.page_size = page_size
         # The request ID.
         self.request_id = request_id
+        self.skills = skills
         # The total number of entries returned. By default, this parameter is not returned.
         self.total_count = total_count
 
     def validate(self):
         if self.data:
             for v1 in self.data:
+                 if v1:
+                    v1.validate()
+        if self.skills:
+            for v1 in self.skills:
                  if v1:
                     v1.validate()
 
@@ -51,6 +57,11 @@ class ListCustomAgentResponseBody(DaraModel):
 
         if self.request_id is not None:
             result['RequestId'] = self.request_id
+
+        result['Skills'] = []
+        if self.skills is not None:
+            for k1 in self.skills:
+                result['Skills'].append(k1.to_map() if k1 else None)
 
         if self.total_count is not None:
             result['TotalCount'] = self.total_count
@@ -74,8 +85,65 @@ class ListCustomAgentResponseBody(DaraModel):
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
 
+        self.skills = []
+        if m.get('Skills') is not None:
+            for k1 in m.get('Skills'):
+                temp_model = main_models.ListCustomAgentResponseBodySkills()
+                self.skills.append(temp_model.from_map(k1))
+
         if m.get('TotalCount') is not None:
             self.total_count = m.get('TotalCount')
+
+        return self
+
+class ListCustomAgentResponseBodySkills(DaraModel):
+    def __init__(
+        self,
+        description: str = None,
+        id: str = None,
+        name: str = None,
+        skill_type: str = None,
+    ):
+        self.description = description
+        self.id = id
+        self.name = name
+        self.skill_type = skill_type
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.description is not None:
+            result['Description'] = self.description
+
+        if self.id is not None:
+            result['Id'] = self.id
+
+        if self.name is not None:
+            result['Name'] = self.name
+
+        if self.skill_type is not None:
+            result['SkillType'] = self.skill_type
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Description') is not None:
+            self.description = m.get('Description')
+
+        if m.get('Id') is not None:
+            self.id = m.get('Id')
+
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+
+        if m.get('SkillType') is not None:
+            self.skill_type = m.get('SkillType')
 
         return self