+++ rdsai-20250507/alibabacloud_rdsai20250507/models/_update_custom_agent_shrink_request.py
@@ -10,6 +10,7 @@ class UpdateCustomAgentShrinkRequest(DaraModel):
         custom_agent_id: str = None,
         enable_tools: bool = None,
         name: str = None,
+        skill_ids_shrink: str = None,
         system_prompt: str = None,
         tools_shrink: str = None,
     ):
@@ -21,6 +22,7 @@ class UpdateCustomAgentShrinkRequest(DaraModel):
         self.enable_tools = enable_tools
         # The ID of the agent.
         self.name = name
+        self.skill_ids_shrink = skill_ids_shrink
         # The name of the agent.
         self.system_prompt = system_prompt
         # Specifies whether to enable tools.
@@ -43,6 +45,9 @@ class UpdateCustomAgentShrinkRequest(DaraModel):
         if self.name is not None:
             result['Name'] = self.name
 
+        if self.skill_ids_shrink is not None:
+            result['SkillIds'] = self.skill_ids_shrink
+
         if self.system_prompt is not None:
             result['SystemPrompt'] = self.system_prompt
 
@@ -61,6 +66,9 @@ class UpdateCustomAgentShrinkRequest(DaraModel):
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
+
+        if m.get('SkillIds') is not None:
+            self.skill_ids_shrink = m.get('SkillIds')
 
         if m.get('SystemPrompt') is not None:
             self.system_prompt = m.get('SystemPrompt')